argument dimid in ncmpii_elem_NC_dimarray() should be of type int
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@1725 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ src/lib/macro.h
@@ -122,8 +122,8 @@ void  NCI_Free_fn(void *ptr, int lineno, const char *func, const char *fname);
                                                                               \
     for (_i=0; _i<_ndims; _i++) {                                             \
         NC_dim *dimp;                                                         \
-	size_t _dimid;                                                        \
-	_dimid = varp->num_subfiles>1?varp->dimids_org[_i]:varp->dimids[_i];  \
+        int _dimid;                                                           \
+        _dimid = (varp->num_subfiles>1)?varp->dimids_org[_i]:varp->dimids[_i];\
         dimp = ncmpii_elem_NC_dimarray(&ncp->dims, _dimid);                   \
         if (dimp->size == NC_UNLIMITED)                                       \
             count[_i] = NC_get_numrecs(ncp);                                  \
@@ -140,7 +140,7 @@ void  NCI_Free_fn(void *ptr, int lineno, const char *func, const char *fname);
                                                                               \
     for (_i=0; _i<varp->ndims; _i++) {                                        \
         NC_dim *dimp;                                                         \
-        dimp = ncmpii_elem_NC_dimarray(&ncp->dims, (size_t)varp->dimids[_i]); \
+        dimp = ncmpii_elem_NC_dimarray(&ncp->dims, varp->dimids[_i]);         \
         if (dimp->size == NC_UNLIMITED)                                       \
             count[_i] = NC_get_numrecs(ncp);                                  \
         else                                                                  \
